days051-060/day059/Harrisons-Blog/requirements.txt
certifi>=2020.6.20
chardet>=3.0.4
click>=7.1.2
dominate>=2.5.2
Flask>=1.1.2
Flask-Bootstrap>=3.3.7.1
Flask-CKEditor>=0.4.4.1
Flask-SQLAlchemy>=2.4.4
Flask-WTF>=0.14.3
idna>=2.10
itsdangerous>=1.1.0
Jinja2>=2.11.2
MarkupSafe>=1.1.1
requests>=2.24.0
SQLAlchemy>=1.3.19
urllib3>=1.25.10
visitor>=0.1.3
Werkzeug>=1.0.1
WTForms>=2.3.3
python-dotenv
flask-login
Flask-Gravatar